+++ packages/generator/vue/src/index.ts
@@ -1,3 +1,4 @@
 export { toEntry as toEntryVue26, toProgram as toProgramVue26 } from "./vue26.js";
 export { toEntry as toEntryVue27, toProgram as toProgramVue27 } from "./vue27.js";
 export { toEntry as toEntryVue3, toProgram as toProgramVue3 } from "./vue3.js";
+// export { toEntry as toEntryVue3, toProgram as toProgramVue3 } from "./vue3-template.js";