+++ packages/components/vue27/CHANGELOG.md
@@ -15,7 +15,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-* **lit-practice-wc/lit-practice-react/lit-practice-vue3/lit-practice-vue27/lit-practice-vue26:** add text Component ([c263b38](https://github.com/sterashima78/lit-practice/commit/c263b389440f3b3bbcdfd4acb4b59b397084221e))
+- **lit-practice-wc/lit-practice-react/lit-practice-vue3/lit-practice-vue27/lit-practice-vue26:** add text Component ([c263b38](https://github.com/sterashima78/lit-practice/commit/c263b389440f3b3bbcdfd4acb4b59b397084221e))
 
 ## [0.0.2](https://github.com/sterashima78/lit-practice/compare/v0.0.2-alpha.1...v0.0.2) (2022-08-18)